requirements.txt
Flask
boto3
PyJWT
python-dotenv
